Move sidebar to component
